pyproject.toml
[project]
name = "xml-data-importer"
version = "0.1.0"
description = "1C data to mariadb importer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "mariadb>=1.1.14",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 120
src = ["importer"]
extend-exclude = [".env"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "SIM", "RET", "Q"]

[tool.ruff.lint.isort]
default-section = "third-party"
combine-as-imports = true
known-local-folder = ["importer"]
length-sort-straight = true
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
